+++ Holovibes/includes/enum/enum_window_kind.hh
@@ -18,25 +18,15 @@ enum class WindowKind
     Filter2D,   /*!< Filter2D view */
 };
 
-const std::map<WindowKind, std::string> window_kind_to_string{{
-                                                                             WindowKind::XYview,
-                                                                             "XYview",
-                                                                         },
-                                                                         {
-                                                                             WindowKind::XZview,
-                                                                             "XZview",
-                                                                         },
-                                                                         {
-                                                                             WindowKind::YZview,
-                                                                             "YZview",
-                                                                         },
-                                                                         {
-                                                                             WindowKind::Filter2D,
-                                                                             "Filter2D",
-                                                                         }};
+// for some reasons clang-format goes bruh here
+// clang-format off
+const std::map<WindowKind, std::string> window_kind_to_string{
+    { WindowKind::XYview, "XYview", },
+    { WindowKind::XZview, "XZview", },
+    { WindowKind::YZview, "YZview", },
+    { WindowKind::Filter2D, "Filter2D", }
+};
+// clang-format on
 
-
-inline std::ostream& operator<<(std::ostream& os, WindowKind obj) {
-    return os << window_kind_to_string.at(obj);
-}
+inline std::ostream& operator<<(std::ostream& os, WindowKind obj) { return os << window_kind_to_string.at(obj); }
 } // namespace holovibes