+++ Holovibes/includes/io_files/output_avi_file.hh
@@ -22,8 +22,8 @@ class OutputAviFile : public OutputFrameFile, public AviFile
 
     /*! \brief Export the compute settings in the file
      *
-     * \param cd The ComputeDescriptor containing the compute settings
-     * \param record_raw Is the raw record enabled
+     * \param input_fps
+     * \param contiguous
      */
     void export_compute_settings(int input_fps, size_t contiguous) override;