+++ Holovibes/sources/cli.cc
@@ -38,7 +38,7 @@ static void progress_bar(int current, int total, int length)
     std::cout.flush();
 }
 
-static void print_verbose(const holovibes::OptionsDescriptor& opts, const holovibes::ComputeDescriptor& cd)
+static void print_verbose(const holovibes::OptionsDescriptor& opts)
 {
     std::cout << "Config:\n";
     boost::property_tree::ptree ptree;
@@ -58,9 +58,7 @@ static void print_verbose(const holovibes::OptionsDescriptor& opts, const holovi
     std::cout << std::endl;
 }
 
-bool get_first_and_last_frame(const holovibes::OptionsDescriptor& opts,
-                              const uint& nb_frames,
-                              holovibes::ComputeDescriptor& cd)
+bool get_first_and_last_frame(const holovibes::OptionsDescriptor& opts, const uint& nb_frames)
 {
     auto err_message = [&](const std::string& name, const uint& value, const std::string& option)
     {
@@ -96,8 +94,6 @@ bool get_first_and_last_frame(const holovibes::OptionsDescriptor& opts,
 
 static bool set_parameters(holovibes::Holovibes& holovibes, const holovibes::OptionsDescriptor& opts)
 {
-    auto& cd = holovibes.get_cd();
-
     std::string input_path = opts.input_path.value();
 
     holovibes::io_files::InputFrameFile* input_frame_file =
@@ -111,7 +107,7 @@ static bool set_parameters(holovibes::Holovibes& holovibes, const holovibes::Opt
 
     const camera::FrameDescriptor& fd = input_frame_file->get_frame_descriptor();
 
-    if (!get_first_and_last_frame(opts, static_cast<uint>(input_frame_file->get_total_nb_frames()), cd))
+    if (!get_first_and_last_frame(opts, static_cast<uint>(input_frame_file->get_total_nb_frames())))
         return false;
 
     holovibes.init_input_queue(fd, holovibes::api::get_input_buffer_size());
@@ -183,8 +179,6 @@ static void main_loop(holovibes::Holovibes& holovibes)
 
 int start_cli(holovibes::Holovibes& holovibes, const holovibes::OptionsDescriptor& opts)
 {
-    auto& cd = holovibes.get_cd();
-
     if (opts.compute_settings_path)
     {
         try
@@ -229,7 +223,7 @@ int start_cli(holovibes::Holovibes& holovibes, const holovibes::OptionsDescripto
 
     if (opts.verbose)
     {
-        print_verbose(opts, cd);
+        print_verbose(opts);
     }
 
     main_loop(holovibes);