+++ Holovibes/sources/gui/windows/MainWindow.cc
@@ -80,9 +80,8 @@ MainWindow::MainWindow(QWidget* parent)
 
     setWindowIcon(QIcon(":/Holovibes.ico"));
 
-    ::holovibes::worker::InformationWorker::display_info_text_function_ = [=](const std::string& text) {
-        synchronize_thread([=]() { ui_->InfoPanel->set_text(text.c_str()); });
-    };
+    ::holovibes::worker::InformationWorker::display_info_text_function_ = [=](const std::string& text)
+    { synchronize_thread([=]() { ui_->InfoPanel->set_text(text.c_str()); }); };
 
     QRect rec = QGuiApplication::primaryScreen()->geometry();
     int screen_height = rec.height();
@@ -158,6 +157,12 @@ MainWindow::MainWindow(QWidget* parent)
     api::start_information_display();
 
     qApp->setStyle(QStyleFactory::create("Fusion"));
+
+    GSH::instance().set_update_view_callback(
+        [&](WindowKind kind, View_Window window)
+        { synchronize_thread([&]() { ui_->ViewPanel->view_callback(kind, window); }); });
+
+    GSH::instance().set_notify_callback([&]() { notify(); });
 }
 
 MainWindow::~MainWindow()
@@ -229,7 +234,8 @@ void MainWindow::notify_error(const std::exception& e)
         const UpdateException* err_update_ptr = dynamic_cast<const UpdateException*>(err_ptr);
         if (err_update_ptr)
         {
-            auto lambda = [&, this] {
+            auto lambda = [&, this]
+            {
                 // notify will be in close_critical_compute
                 api::handle_update_exception();
                 api::close_windows();
@@ -241,7 +247,8 @@ void MainWindow::notify_error(const std::exception& e)
             synchronize_thread(lambda);
         }
 
-        auto lambda = [&, this, accu = (dynamic_cast<const AccumulationException*>(err_ptr) != nullptr)] {
+        auto lambda = [&, this, accu = (dynamic_cast<const AccumulationException*>(err_ptr) != nullptr)]
+        {
             if (accu)
             {
                 handle_accumulation_exception();
@@ -263,11 +270,13 @@ void MainWindow::notify_error(const std::exception& e)
 
 void MainWindow::layout_toggled()
 {
-    synchronize_thread([=]() {
-        // Resizing to original size, then adjust it to fit the groupboxes
-        resize(baseSize());
-        adjustSize();
-    });
+    synchronize_thread(
+        [=]()
+        {
+            // Resizing to original size, then adjust it to fit the groupboxes
+            resize(baseSize());
+            adjustSize();
+        });
 }
 
 void MainWindow::credits()